Navigator tool accepts arrow keys
Listen for arrow key input and make that reflect in the current location. This involves posting an event that the UI can process.
Each arrow key becomes a "nudge" that increases / decreases the axis value with 1 unit.
This may be problematic for axes that are between -1, 0, 1 or 0, 1 but heck these will have problems anyway.